requirements-dev.txt
ruff==0.14.14
pytest==9.0.2
pytest-cov==7.0.0

Flask==3.1.2
flask-sanitize-escape==0.0.3
gunicorn==24.1.1
mysql-connector-python==9.5.0
python-slugify==8.0.4
pytz==2025.2
